+++ client/python/vstimd/system/__init__.py
@@ -1,5 +1,6 @@
 from .system_client import SystemClient
 from .system_models import (
+    CapturedFrame,
     DeferredModeStatus,
     ServerInfo,
     ServerVersion,
@@ -8,6 +9,7 @@ from .system_models import (
 
 __all__ = [
     "SystemClient",
+    "CapturedFrame",
     "DeferredModeStatus",
     "ServerInfo",
     "ServerVersion",